+++ etils/ecolab/__init__.py
@@ -14,5 +14,7 @@
 
 """Colab public API."""
 
+
 from etils.ecolab.array_as_img import auto_plot_array
 from etils.ecolab.colab_utils import collapse
+from etils.ecolab.lazy_imports import print_lazy_imports